Aggregate ordering fix
Alter 5 to not rebuild geometry is optimised to 3 level

+++ rifDatabase/Postgres/psql_scripts/alter_scripts/v4_0_alter_5.sql
@@ -515,6 +515,11 @@ COMMENT ON TYPE rif40_goejson_type IS 'Special type to allow ROW() elements to b
 	
 DO LANGUAGE plpgsql $$
 DECLARE
+	c1alter5 CURSOR FOR
+		SELECT column_name 
+		  FROM information_schema.columns 
+		 WHERE table_name = 't_rif40_sahsu_geometry' AND column_name = 'optimised_geojson_3' AND table_schema = 'rif_data';
+	c1_rec RECORD;
 --
 	rif40_geo_pkg_functions 	VARCHAR[] := ARRAY['lf_check_rif40_hierarchy_lookup_tables', 
 							'populate_rif40_geometry_tables', 
@@ -541,8 +546,8 @@ BEGIN
 --
 -- Turn on some debug
 --
-        PERFORM rif40_log_pkg.rif40_log_setup();
-        PERFORM rif40_log_pkg.rif40_send_debug_to_info(TRUE);
+    PERFORM rif40_log_pkg.rif40_log_setup();
+    PERFORM rif40_log_pkg.rif40_send_debug_to_info(TRUE);
 --
 -- Enabled debug on select rif40_sm_pkg functions
 --
@@ -553,25 +558,34 @@ BEGIN
 
 	END LOOP;
 --
+-- Check if run already
+--
+	OPEN c1alter5;
+	FETCH c1alter5 INTO c1_rec;
+	CLOSE c1alter5;
+	IF c1_rec.column_name = 'optimised_geojson_3' THEN
+		RAISE INFO 'Column: t_rif40_sahsu_geometry.optimised_geojson_3 exists; no need to rebuild geometry tables';
+	ELSE
+--
 -- Drop old geometry tables
 --
-	PERFORM rif40_geo_pkg.drop_rif40_geolevels_geometry_tables('SAHSU');
-	PERFORM rif40_geo_pkg.drop_rif40_geolevels_lookup_tables('SAHSU');
-	DROP TABLE IF EXISTS sahsuland_geography_orig;
+		PERFORM rif40_geo_pkg.drop_rif40_geolevels_geometry_tables('SAHSU');
+		PERFORM rif40_geo_pkg.drop_rif40_geolevels_lookup_tables('SAHSU');
+		DROP TABLE IF EXISTS sahsuland_geography_orig;
 --
 -- These are the new T_RIF40_<GEOGRAPHY>_GEOMETRY tables and
 -- new p_rif40_geolevels_geometry_<GEOGRAPHY>_<GEOELVELS> partitioned tables
 --
-	PERFORM rif40_geo_pkg.create_rif40_geolevels_geometry_tables('SAHSU');
+		PERFORM rif40_geo_pkg.create_rif40_geolevels_geometry_tables('SAHSU');
 --	RAISE EXCEPTION 'Stop!!!';
 --
 -- Create and populate rif40_geolevels lookup and create hierarchy tables 
 --
-	PERFORM rif40_geo_pkg.create_rif40_geolevels_lookup_tables('SAHSU');
+		PERFORM rif40_geo_pkg.create_rif40_geolevels_lookup_tables('SAHSU');
 --
 -- Populate geometry tables
 --
-	PERFORM rif40_geo_pkg.populate_rif40_geometry_tables('SAHSU');
+		PERFORM rif40_geo_pkg.populate_rif40_geometry_tables('SAHSU');
 --
 -- Simplify geometry
 --
@@ -579,7 +593,7 @@ BEGIN
 --
 -- psql:rif40_geolevels_ew01_geometry.sql:174: ERROR:  Error performing intersection: TopologyException: found non-noded intersection between LINESTRING (-2.9938 53.3669, -2.98342 53.367) and LINESTRING (-2.98556 53.367, -2.98556 53.367) at -2.9855578257498334 53.366966593247653
 --
-	PERFORM rif40_geo_pkg.simplify_geometry('SAHSU', 10 /* l_min_point_resolution [1] */);
+		PERFORM rif40_geo_pkg.simplify_geometry('SAHSU', 10 /* l_min_point_resolution [1] */);
 --
 -- Populate hierarchy tables
 --
@@ -594,29 +608,30 @@ BEGIN
 --  FROM p_rif40_geolevels_geometry_sahsu_level3 a3, p_rif40_geolevels_geometry_sahsu_level2 a2  
 -- WHERE ST_Intersects(a2.optimised_geometry, a3.optimised_geometry);
 --
-	PERFORM rif40_geo_pkg.populate_hierarchy_table('SAHSU'); 
+		PERFORM rif40_geo_pkg.populate_hierarchy_table('SAHSU'); 
 --
 -- Add denominator population table to geography geolevel geomtry data
 --
-	PERFORM rif40_geo_pkg.add_population_to_rif40_geolevels_geometry('SAHSU', 'SAHSULAND_POP'); 
+		PERFORM rif40_geo_pkg.add_population_to_rif40_geolevels_geometry('SAHSU', 'SAHSULAND_POP'); 
 --
 -- Fix NULL geolevel names in geography geolevel geometry and lookup table data 
 --
-	PERFORM rif40_geo_pkg.fix_null_geolevel_names('SAHSU'); 
+		PERFORM rif40_geo_pkg.fix_null_geolevel_names('SAHSU'); 
 --
 -- Make level1 names consistent
 --
-	UPDATE sahsuland_level1 a 
-	   SET name = (
-		SELECT name 
-		  FROM t_rif40_sahsu_geometry b
-		 WHERE b.geolevel_name = 'LEVEL1'
-		   AND b.area_id = a.level1);
+		UPDATE sahsuland_level1 a 
+		SET name = (
+			SELECT name 
+			  FROM t_rif40_sahsu_geometry b
+ 			 WHERE b.geolevel_name = 'LEVEL1'
+		       AND b.area_id = a.level1);
 --
 -- Add: gid_rowindex (i.e 1_1). Where gid corresponds to gid in geometry table
 -- row_index is an incremental serial aggregated by gid ( starts from one for each gid)
 --
-	PERFORM rif40_geo_pkg.gid_rowindex_fix('SAHSU');
+		PERFORM rif40_geo_pkg.gid_rowindex_fix('SAHSU');
+	END IF;
 --
 -- Populate Map tiles
 --
